Update : Added Like and Delete Layout

+++ assignment/Screens/Favorite/View/FavoriteViewController.swift
@@ -11,6 +11,7 @@ import SDWebImage
 class FavoriteViewController: UIViewController, UIGestureRecognizerDelegate {
     @Published private var models=[FavoriteItems]()
     private var longPressGesture:UILongPressGestureRecognizer!
+    private var tapGestures:UITapGestureRecognizer!
     
     @IBOutlet weak var favoriteCollectionView: UICollectionView!
     
@@ -19,7 +20,17 @@ class FavoriteViewController: UIViewController, UIGestureRecognizerDelegate {
         models=DbService.shareInstance.getAllItem()
         favoriteCollectionView.reloadData()
         
+        
 
+    }
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        models=DbService.shareInstance.getAllItem()
+        DispatchQueue.main.async {
+            self.favoriteCollectionView.reloadData()
+
+        }
+        
     }
    
 }
@@ -32,7 +43,6 @@ extension FavoriteViewController:UICollectionViewDelegate,UICollectionViewDataSo
         let favCell=(favoriteCollectionView.dequeueReusableCell(withReuseIdentifier: "favCell", for: indexPath) as? FavoriteCollectionViewCell)!
         
         let model=models[indexPath.row]
-        print(model)
         guard let imageURL=model.url else{
             favCell.favoriteImageView.image=UIImage(imageLiteralResourceName: "emptyImage")
             return favCell;
@@ -43,26 +53,50 @@ extension FavoriteViewController:UICollectionViewDelegate,UICollectionViewDataSo
         favCell.favoriteImageView.contentMode = .scaleAspectFill
         favCell.favoriteImageView.layer.cornerRadius=5
         
-        longPressGesture = UILongPressGestureRecognizer(target: self, action: #selector(longPress(_:)))
-        longPressGesture.minimumPressDuration = 0.5
-        longPressGesture.delegate = self
-        longPressGesture.delaysTouchesBegan = true
-       favCell.addGestureRecognizer(longPressGesture)
+//        longPressGesture = UILongPressGestureRecognizer(target: self, action: #selector(longPress(_:)))
+//        longPressGesture.minimumPressDuration = 2
+//        longPressGesture.delaysTouchesBegan = true
+        tapGestures = UITapGestureRecognizer(target: self, action: #selector(didTripleTap(_:)))
+        tapGestures.numberOfTapsRequired = 3
+        favCell.addGestureRecognizer(tapGestures)
+      
         return favCell
     }
-    @objc func longPress(_ gesture: UILongPressGestureRecognizer) {
+    @objc func didTripleTap(_ gesture: UITapGestureRecognizer) {
+        deleteButton(gesture)
         let tap = gesture.location(in: self.favoriteCollectionView)
-        let index:NSIndexPath=self.favoriteCollectionView.indexPathForItem(at: tap)! as NSIndexPath
-        
-        print("Long Pressed : ",index.row)
+        guard let index:NSIndexPath=self.favoriteCollectionView.indexPathForItem(at: tap) as? NSIndexPath else{
+            return
+        }
+        print("Triple Tap : ",index.row)
         let item = models[index.row]
+        DispatchQueue.main.asyncAfter(deadline: .now()+0.8){
+            DbService.shareInstance.deleteItem(item: item)
+            self.models=DbService.shareInstance.getAllItem()
+            self.favoriteCollectionView.reloadData()
+           }
        
-        DbService.shareInstance.deleteItem(item: item)
-        self.favoriteCollectionView.reloadData()
-        
         
         }
     
-    
+    func deleteButton(_ gesture: UITapGestureRecognizer){
+        let trash=UIImageView(image: UIImage(systemName: "trash.fill"))
+        gesture.view?.addSubview(trash)
+        let size=gesture.view!.frame.size.width/4
+        trash.tintColor = .red
+        trash.frame=CGRect(x: ((gesture.view?.frame.size.height)!)/3, y: ((gesture.view?.frame.size.width)!)/4, width: size, height: size)
+        
+        trash.center = CGPoint(x: gesture.view!.frame.size.width  / 2,
+                               y: gesture.view!.frame.size.height / 2)
+        DispatchQueue.main.asyncAfter(deadline: .now()+0.3, execute: {
+            UIView.animate(withDuration: 0.5, animations: {
+                trash.alpha=0
+            },completion: {done in
+                if done{
+                    trash.removeFromSuperview()
+                }
+            })
+        })
+    }
     
 }
